All planets closer to COM than planet
This version finds every planet which is closer to the COM than the planet selected for every planet in the system. This is a larger set than the answer. Need to remove the planets closer to COM which are not not in a direct orbit chain to the COM. 

that is if a)b)c)e
and b)d
e is not an indirect orbit of d.